+++ CHANGES.rst
@@ -23,6 +23,11 @@ esa.hubble
 - Status and maintenance messages from eHST TAP when the module is instantiated. get_status_messages method to retrieve them. [#2597]
 - Optional parameters in all methods are kwargs keyword only. [#2597]
 
+solarsystem.neodys
+^^^^^^^^^^^^^^^^^^
+
+- New module to access the NEODyS web interface.
+
 solarsystem.pds
 ^^^^^^^^^^^^^^^